refactor(weather-service): extract rate limiting to an additional package
