+++ weather-service/ratelimit/ratelimit.go
@@ -1,0 +1,59 @@
+package ratelimit
+
+import (
+	"encoding/json"
+	"net/http"
+	"strings"
+	"sync"
+
+	"golang.org/x/time/rate"
+)
+
+// IPRateLimiter Create a custom limiter structure
+type IPRateLimiter struct {
+	ips map[string]*rate.Limiter
+	mu  *sync.RWMutex
+	r   rate.Limit
+	b   int
+}
+
+// NewIPRateLimiter Create a new rate limiter that allows x requests per second with burst of y
+func NewIPRateLimiter(r rate.Limit, b int) *IPRateLimiter {
+	return &IPRateLimiter{
+		ips: make(map[string]*rate.Limiter),
+		mu:  &sync.RWMutex{},
+		r:   r,
+		b:   b,
+	}
+}
+
+// GetLimiter Get or create a limiter for an IP address
+func (i *IPRateLimiter) GetLimiter(ip string) *rate.Limiter {
+	i.mu.Lock()
+	defer i.mu.Unlock()
+
+	limiter, exists := i.ips[ip]
+	if !exists {
+		limiter = rate.NewLimiter(i.r, i.b)
+		i.ips[ip] = limiter
+	}
+
+	return limiter
+}
+
+// Middleware creates a new middleware handler for rate limiting
+func Middleware(next http.HandlerFunc, limiter *IPRateLimiter) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		ip := strings.Split(r.RemoteAddr, ":")[0]
+		if !limiter.GetLimiter(ip).Allow() {
+			w.Header().Set("Content-Type", "application/json")
+			w.WriteHeader(http.StatusTooManyRequests)
+			json.NewEncoder(w).Encode(map[string]string{
+				"error":   "Rate limit exceeded",
+				"message": "Too many requests, please try again later",
+			})
+			return
+		}
+		next.ServeHTTP(w, r)
+	}
+}
